+++ CHANGELOG.md
@@ -58,7 +58,7 @@ And please only add new entries to the top of this list, right below the `# Unre
 - **Breaking:** On X11, device events are now ignored for unfocused windows by default, use `EventLoopWindowTarget::set_device_event_filter` to set the filter level.
 - Implemented `Default` on `EventLoop<()>`.
 - Implemented `Eq` for `Fullscreen`, `Theme`, and `UserAttentionType`.
-- **Breaking:** `Window::set_cursor_grab` got renamed to `Window::set_cursor_grab_mode` and now accepts `CursorGrabMode` to control grabbing behavior.
+- **Breaking:** `Window::set_cursor_grab` now accepts `CursorGrabMode` to control grabbing behavior.
 - On Wayland, add support for `Window::set_cursor_position`.
 
 # 0.26.1 (2022-01-05)